requirements.txt
asgiref==3.8.1
Django==4.2.11
django-ckeditor==6.7.1
django-environ==0.11.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
